Audit the session's code: fix a calibration bug that could move the rank guards
Audited everything changed this session against live code rather than by
re-reading it. Most claims held; one did not.

THE BUG. Platt is monotone, so it can never REORDER cells -- but it can MERGE
them, and a merged pair changes a rank metric. Two mechanisms, both measured:

  * the EPS clamp at 1e-6 collapsed 4 of 8 float32-representable values near 1.0
    onto one number. float32's ulp there is ~6e-08, so the clamp was coarser than
    the input's own resolution.
  * casting calibrated probabilities back to float32 in _summaries merged values
    the map had left distinct. At slope 20 that alone moved PR-AUC 1.000 -> 0.941
    and ROC-AUC -> 0.966.

_PLATT_BOUNDS permits a slope up to 1e2, so a steep fit is reachable on real
data. Had one occurred, fate_prauc would have reported a REGRESSION -- a Stage 1
GUARD -- and the correct response under section 3 is to revert. We would have
reverted a working change because of a rounding artefact.

Fixed: EPS -> 1e-9 (two orders below the float32 ulp, so every representable
input except exact 0/1 stays distinct), a numerically stable sigmoid for the
wider logit range that admits, and _summaries no longer downcasts -- _rows
converts to Python floats and res.py upcasts to float64, so nothing downstream
wanted the narrower type. Guards now hold at slopes 2, 8, 20 and 100, pinned by
test_calibration_does_not_move_the_rank_guards_even_at_a_steep_slope.

CLAIMS CORRECTED. Four places said Platt makes the rank guards "mathematically
invariant" or "bit-identical". Too strong: monotone means no reordering, not no
merging. Corrected in CHANGES.md, the lab notebook, smoke_stage1.py and
common/calibration.py.

VERIFIED RATHER THAN ASSUMED
  * biology untouched: the diff over src/cellfate/data, models and evaluation is
    empty -- clock, harmonization, fate labels, dAge targets, network unchanged.
  * column binding: indices 0-5 unchanged, donor never reaches forward().
  * ROW ALIGNMENT: rebuilt a dataset and compared every donor code against the
    shard's cell_line -- 144/144 match.
  * Platt recovers a 3x sharpening, a +1.8 bias, and both, to within 0.02 of the
    true inverse.
  * simplex invariants hold on saturated/zero/uniform input; loss:death ratio
    preserved; schema rejects a negative slope and a half-specified pair; legacy
    bundles load unchanged; xstats round-trips.

SCOPE CHECK ON REAL BUNDLES. Retrained the six rehearsal folds and compared
against the same folds trained before Change A": conformal_q and sigma_scale are
BIT-IDENTICAL (0.47744181752204895 and 7.795770789209797 on N2), temperature
1.498 -> 1.0 as Platt replaces it. The calibrator provably does not reach q.

HELD-OUT COMPARISON (synthetic, 3 folds x 10 cells -- weak, directional only):
cross-donor temperature is the WORST option at 0.190 mean ECE, independently
reproducing run 2's regression on data it was never fitted to; the shipped union
Platt is best at 0.153. The synthetic setup does not reproduce the real
miscalibration magnitude, so this is support for the direction, not a prediction
that run 3 clears the bar.

221 tests pass; smoke 34/34.

+++ src/cellfate/inference/predictor.py
@@ -160,9 +160,16 @@ class Predictor:
         # calibrator was fitted on (xstats.probs_mean). Fitting on mean-logits and applying
         # per-member -- what temperature does -- calibrates a different thing by Jensen.
         if self.platt is not None:
+            # Kept in float64, NOT cast back to the members' float32. Platt with a large slope
+            # saturates confident predictions toward 1.0, and a float32 cast then merges values
+            # the map left distinct -- creating TIES. Measured at slope 20 on overlapping
+            # classes: PR-AUC 1.000 -> 0.941 and ROC-AUC -> 0.966 purely from that rounding,
+            # which would read as a Stage 1 GUARD REGRESSION and trigger a wrongful revert.
+            # `_rows` converts to Python floats anyway and `res.py` upcasts to float64, so
+            # nothing downstream wants the narrower type.
             pbar = torch.from_numpy(
                 apply_platt(pbar.cpu().numpy(), *self.platt, safe_idx=SAFE_IDX)
-            ).to(pbar.dtype)
+            )
         ent = -(pbar * (pbar + 1e-12).log()).sum(1)
         Z = z.detach().cpu().numpy()
         return {
